Add Ui.java and Parser.java
