--- data/taskalyn.txt
@@ -1,1 +1,0 @@
-D | 1 | synopsis | 2024-08-30 1400
