+++ src/main/java/Ui.java
@@ -1,0 +1,28 @@
+import java.util.Scanner;
+
+public class Ui {
+    private Scanner scanner;
+
+    public Ui() {
+        scanner = new Scanner(System.in);
+    }
+
+    public void printLines(String input) {
+        System.out.println("    ____________________________________________________________");
+        System.out.println("    " + input);
+        System.out.println("    ____________________________________________________________\n");
+    }
+
+    public void printWelcome() {
+        printLines("Hey! I'm Taskalyn, your personal Task Manager :)\n" +
+                "    What can I do for you?");
+    }
+
+    public String readCommand() {
+        return scanner.nextLine().trim();
+    }
+
+    public void close() {
+        scanner.close();
+    }
+}
